Modify login user id
Not use MongoDB auto-added _id field since the server cannot identify
users that have signed in before in the sign-in process. And that will
cause duplicated data for the same user.
Instead, use "prefix + email" as user id. The prefix string can be
"google-" for those users signing in with their Google account. If the
web app will support other login ways, e.g. Facebook or LinkedIn, in the
future, the prefix string remains the flexibility of expansion.

other changes:
- add logout function
- add session control.
  user don't need to login again until session expired.